S3: put_object() - IfMatch is now supported
