+++ tests/test_s3/test_s3.py
@@ -20,7 +20,6 @@ from freezegun import freeze_time
 
 import moto.s3.models as s3model
 from moto import mock_aws, moto_proxy, settings
-from moto.core.utils import utcnow
 from moto.moto_api import state_manager
 from moto.s3.models import s3_backends
 from moto.s3.responses import DEFAULT_REGION_NAME
@@ -524,37 +523,6 @@ def test_key_with_special_characters(key):
 
     resp = client.get_object(Bucket="testname", Key=key)
     assert resp["Body"].read() == b"value"
-
-
-@s3_aws_verified
-@pytest.mark.aws_verified
-@pytest.mark.parametrize("versioned", [True, False], ids=["versioned", "not versioned"])
-def test_conditional_write(versioned, bucket_name=None):
-    s3 = boto3.client("s3", region_name="us-east-1")
-    if versioned:
-        enable_versioning(bucket_name, s3)
-
-    s3.put_object(
-        Key="test_object",
-        Body="test",
-        Bucket=bucket_name,
-        IfNoneMatch="*",
-    )
-
-    with pytest.raises(ClientError) as exc:
-        s3.put_object(
-            Key="test_object",
-            Body="another_test",
-            Bucket=bucket_name,
-            IfNoneMatch="*",
-        )
-    err = exc.value.response["Error"]
-    assert err["Code"] == "PreconditionFailed"
-    assert (
-        err["Message"]
-        == "At least one of the pre-conditions you specified did not hold"
-    )
-    assert err["Condition"] == "If-None-Match"
 
 
 @mock_aws
@@ -2008,204 +1976,6 @@ def test_delete_versioned_bucket_returns_metadata(bucket_name=None):
     )
     assert "DeleteMarker" not in del_mrk4
     assert del_mrk4["VersionId"] == del_mrk2["VersionId"]
-
-
-@mock_aws
-def test_get_object_if_modified_since_refresh():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    response = s3_client.get_object(Bucket=bucket_name, Key=key)
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.get_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfModifiedSince=response["LastModified"],
-        )
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-
-
-@mock_aws
-def test_get_object_if_modified_since():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.get_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfModifiedSince=utcnow() + datetime.timedelta(hours=1),
-        )
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-
-
-@mock_aws
-def test_get_object_if_unmodified_since():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.get_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfUnmodifiedSince=utcnow() - datetime.timedelta(hours=1),
-        )
-    err_value = err.value
-    assert err_value.response["Error"]["Code"] == "PreconditionFailed"
-    assert err_value.response["Error"]["Condition"] == "If-Unmodified-Since"
-
-
-@mock_aws
-def test_get_object_if_match():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.get_object(Bucket=bucket_name, Key=key, IfMatch='"hello"')
-    err_value = err.value
-    assert err_value.response["Error"]["Code"] == "PreconditionFailed"
-    assert err_value.response["Error"]["Condition"] == "If-Match"
-
-
-@mock_aws
-def test_get_object_if_none_match():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    etag = s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")["ETag"]
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.get_object(Bucket=bucket_name, Key=key, IfNoneMatch=etag)
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-    assert err_value.response["ResponseMetadata"]["HTTPHeaders"]["etag"] == etag
-
-
-@mock_aws
-def test_head_object_if_modified_since():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.head_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfModifiedSince=utcnow() + datetime.timedelta(hours=1),
-        )
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-
-
-@mock_aws
-def test_head_object_if_modified_since_refresh():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    response = s3_client.head_object(Bucket=bucket_name, Key=key)
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.head_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfModifiedSince=response["LastModified"],
-        )
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-
-
-@mock_aws
-def test_head_object_if_unmodified_since():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.head_object(
-            Bucket=bucket_name,
-            Key=key,
-            IfUnmodifiedSince=utcnow() - datetime.timedelta(hours=1),
-        )
-    err_value = err.value
-    assert err_value.response["Error"] == {
-        "Code": "412",
-        "Message": "Precondition Failed",
-    }
-
-
-@mock_aws
-def test_head_object_if_match():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.head_object(Bucket=bucket_name, Key=key, IfMatch='"hello"')
-    err_value = err.value
-    assert err_value.response["Error"] == {
-        "Code": "412",
-        "Message": "Precondition Failed",
-    }
-
-
-@mock_aws
-def test_head_object_if_none_match():
-    s3_client = boto3.client("s3", region_name=DEFAULT_REGION_NAME)
-    bucket_name = "blah"
-    s3_client.create_bucket(Bucket=bucket_name)
-
-    key = "hello.txt"
-
-    etag = s3_client.put_object(Bucket=bucket_name, Key=key, Body="test")["ETag"]
-
-    with pytest.raises(botocore.exceptions.ClientError) as err:
-        s3_client.head_object(Bucket=bucket_name, Key=key, IfNoneMatch=etag)
-    err_value = err.value
-    assert err_value.response["Error"] == {"Code": "304", "Message": "Not Modified"}
-    assert err_value.response["ResponseMetadata"]["HTTPHeaders"]["etag"] == etag
 
 
 @mock_aws
